minor docs fix
NOTE: Navigation has become more strict now, if you use the function that takes only one URL, then the navigation will only pass if this URL is the same final URL you end up with and no redirection occurs (not even double slashes or anything else). If anything is added/removed/modified from the target url, the navigation is expected to FAIL. In case you're expecting a redirection to occur, then you can use the overloaded navigation function that accepts a targetURL and a subset of the final url that you're expecting to end up with.